refactor(session): split role lifecycle
