+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Session/Client/Session.ClientTick.cs
@@ -4,6 +4,7 @@ namespace ShipCoreFramework
     {
         private void RunClientSimulationTick()
         {
+            CoreTerminalControls.RegisterOnce();
             CoreTypeLCDScript.RunFrameScrollUpdate();
             NotificationInstance.RunCountdownTick();
         }
