+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Session/Server/Session.ServerFields.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Concurrent;
 using NexusModAPI;
 using VRage.Game.ModAPI;
@@ -12,6 +13,9 @@ namespace ShipCoreFramework
         private bool _startedNexus;
         private int _massCacheRefreshCursor;
 
+        internal static bool HasStarted;
+        internal static readonly Guid CoreStateStorageGUID =
+            new Guid("a8807ad4-524d-441a-a89a-0671fbfb1dd3");
         internal static readonly ConcurrentDictionary<IMyGridGroupData, PhysicalSpeedCluster> PhysicalSpeedClusterDict =
             new ConcurrentDictionary<IMyGridGroupData, PhysicalSpeedCluster>();
     }
