+++ ShipCoreFramework/src/Data/Scripts/ShipCoreFramework/Session/Session.Run.cs
@@ -1,9 +1,5 @@
 using System;
 using System.Collections.Generic;
-using NexusModAPI;
-using Sandbox.Common.ObjectBuilders;
-using Sandbox.Definitions;
-using Sandbox.Game;
 using Sandbox.Game.Entities;
 using Sandbox.ModAPI;
 using VRage.Game;
@@ -15,11 +11,6 @@ namespace ShipCoreFramework
     [MySessionComponentDescriptor(MyUpdateOrder.AfterSimulation, 0)]
     public partial class Session : MySessionComponentBase
     {
-        private static readonly string[] AmmoDefinitionIds =
-        {
-            "Missile", "LargeCalibreShell", "MediumCalibreShell", "LargeCaliber", "AutocannonShell",
-            "LargeRailgunSlug", "SmallRailgunSlug", "SmallCaliber", "PistolCaliber", "Shrapnel"
-        };
         public override void BeforeStart()
         {
             ModAPI.Initialize();
@@ -28,24 +19,18 @@ namespace ShipCoreFramework
             MyAPIGateway.GridGroups.OnGridGroupDestroyed += GridGroupsOnOnGridGroupDestroyed;
             MyCubeGrid.OnBlocksChangeFinishedGlobally += MyCubeGridOnBlocksChangeFinishedGlobally;
             
-            var mechanicalGroups = new List<IMyGridGroupData>();
-            var physicalGroups = new List<IMyGridGroupData>();
-            MyAPIGateway.GridGroups.GetGridGroups(GridLinkTypeEnum.Mechanical, mechanicalGroups);
-            MyAPIGateway.GridGroups.GetGridGroups(GridLinkTypeEnum.Physical, physicalGroups);
-            Utils.Log("BeforeStart: found " + mechanicalGroups.Count + " mechanical groups and " +
-                      physicalGroups.Count + " physical groups for initial scan.", 1);
+            var initialGroups = new List<IMyGridGroupData>();
+            MyAPIGateway.GridGroups.GetGridGroups(GridLinkTypeEnum.Mechanical, initialGroups);
+            if (IsServer)
+                AppendInitialPhysicalGroups(initialGroups);
+            Utils.Log("BeforeStart: found " + initialGroups.Count + " grid groups for initial scan.", 1);
 
             IsInitialGroupScan = true;
             try
             {
-                MyAPIGateway.Parallel.ForEach(mechanicalGroups, mechanicalGroup =>
+                MyAPIGateway.Parallel.ForEach(initialGroups, group =>
                 {
-                    GridGroupsOnOnGridGroupCreated(mechanicalGroup);
-                });
-
-                MyAPIGateway.Parallel.ForEach(physicalGroups, physicalGroup =>
-                {
-                    GridGroupsOnOnGridGroupCreated(physicalGroup);
+                    GridGroupsOnOnGridGroupCreated(group);
                 });
             }
             finally
@@ -58,65 +43,33 @@ namespace ShipCoreFramework
         {
             GameThreadId = Environment.CurrentManagedThreadId;
             IsShuttingDown = false;
+            _tick = 0;
+            CurrentTick = 0;
             MpActive = MyAPIGateway.Multiplayer.MultiplayerActive;
             IsServer = (MpActive && MyAPIGateway.Multiplayer.IsServer) || !MpActive;
             IsClient = (MpActive && !MyAPIGateway.Utilities.IsDedicated) || !MpActive;
 
+            if (Networking == null)
+                Networking = new Networking(32124);
             Networking.Register();
             Config.LoadConfig(IsServer);
             if (IsClient)
-            {
-                try
-                {
-                    ApplyHighResolutionLcdDefinitions();
-                }
-                catch (Exception e)
-                {
-                    Utils.Log("High-resolution LCD setup skipped: " + e.Message, 1);
-                }
-            }
+                LoadClientData();
+            if (IsServer)
+                LoadServerData();
 
             Utils.Log("LoadData: MpActive=" + MpActive + ", IsServer=" + IsServer +
                       ", IsClient=" + IsClient + ", Dedicated=" + MyAPIGateway.Utilities.IsDedicated + ".", 1);
-            _myNexusApi = new NexusAPI(OnNexusEnabled);
-
-            if (IsServer)
-            {
-                ApplyConfigToDefinitions();
-            }
-            else if (MpActive)
+            if (IsClient && !IsServer && MpActive)
             {
                 Networking.SendToServer(new PacketRequestConfig(), onlyToServer: true);
             }
-            else
-            {
-                ApplyConfigToDefinitions();
-            }
-            
-            MyAPIGateway.Session.OnSessionReady += SessionReady;
-            MyAPIGateway.Session.Factions.FactionStateChanged += FactionStateChanged;
-            MyAPIGateway.Session.Factions.FactionCreated += FactionCreated;
-            MyAPIGateway.Session.Factions.FactionEdited += FactionEdited;
-            MyAPIGateway.Utilities.MessageEnteredSender += Commands.OnChatCommand;
-            if(IsServer)
-            {
-                MyAPIGateway.Multiplayer.RegisterSecureMessageHandler(CommandsSyncId, Commands.ServerMessageHandler);
-                Utils.Log("Ship Cores: Awaiting Commands From Clients", 1);
-            }
-            if (IsServer) Config.SaveConfig();
         }
 
         protected override void UnloadData()
         {
             Utils.Log("UnloadData: shutting down Ship Core Framework session.", 1);
             IsShuttingDown = true;
-            MyAPIGateway.Session.OnSessionReady -= SessionReady;
-            MyAPIGateway.Session.Factions.FactionStateChanged -= FactionStateChanged;
-            MyAPIGateway.Session.Factions.FactionCreated -= FactionCreated;
-            MyAPIGateway.Session.Factions.FactionEdited -= FactionEdited;
-            MyAPIGateway.Utilities.MessageEnteredSender -= Commands.OnChatCommand;
-            MyAPIGateway.Multiplayer.UnregisterSecureMessageHandler(CommandsSyncId, Commands.ServerMessageHandler);
-            MyExplosions.OnExplosion -= CubeGridModifiers.HandleLightningExplosions;
             
             try //Because this throws a NRE in keen code if you alt-F4
             {
@@ -126,24 +79,13 @@ namespace ShipCoreFramework
             }
             catch { /**/ }
 
-            UntrackAllPhysicalGridGroups();
+            if (IsClient)
+                UnloadClientData();
+            if (IsServer)
+                UnloadServerData();
 
             RevertAmmoSpeedAdjustments();
-            RevertHighResolutionLcdDefinitions();
-            CoreTerminalControls.Unregister();
-            
-            LimitsNexusSync.Stop();
-            _myNexusApi?.Unload();
-            _myNexusApi = null;
-
             ModAPI.Close();
-            RuntimeStateStore.Clear();
-            ResetRuntimeStateSync();
-
-            PerFactionManager.Reset();
-            PerPlayerManager.Reset();
-            PerManifestGroupManager.Reset();
-            if (IsServer) Config.SaveConfig();
             Networking?.Unregister();
             Networking = null;
             
@@ -160,11 +102,10 @@ namespace ShipCoreFramework
         {
             if (Config.SelectedNoCore == null)
             {
-                Utils.ShowNotification("There is no No Core currently selected. Make sure to select one and reload the world!!", 0, 20000, true);
+                if (IsServer)
+                    NotifyMissingNoCore();
                 return;
             }
-            if(!HasStarted) HasStarted = true;
-            CoreTerminalControls.RegisterOnce();
 
             _tick++;
             CurrentTick = _tick;
@@ -172,71 +113,8 @@ namespace ShipCoreFramework
                 RunClientSimulationTick();
 
             if (!IsServer) return;
+            if (!HasStarted) HasStarted = true;
             RunServerSimulationTick();
-        }
-
-        internal static void ApplyConfigToDefinitions()
-        {
-            if (MyDefinitionManager.Static?.EnvironmentDefinition != null)
-            {
-                MyDefinitionManager.Static.EnvironmentDefinition.LargeShipMaxSpeed = Config.MaxPossibleSpeedMetersPerSecond;
-                MyDefinitionManager.Static.EnvironmentDefinition.SmallShipMaxSpeed = Config.MaxPossibleSpeedMetersPerSecond;
-            }
-
-            var newDifferential = Config.MaxPossibleSpeedMetersPerSecond - 100.0f;
-            var delta = newDifferential - AppliedSpeedDifferential;
-            if (Math.Abs(delta) < 0.001f)
-                return;
-
-            foreach (var ammoId in AmmoDefinitionIds)
-            {
-                try
-                {
-                    var ammoDefinition = MyDefinitionManager.Static?.GetAmmoDefinition(
-                        new MyDefinitionId(typeof(MyObjectBuilder_AmmoDefinition), ammoId));
-                    if (ammoDefinition != null)
-                        ammoDefinition.DesiredSpeed += delta;
-                }
-                catch
-                {
-                    // Ignore missing ammo definitions.
-                }
-            }
-
-            AppliedSpeedDifferential = newDifferential;
-        }
-
-        private static void RevertAmmoSpeedAdjustments()
-        {
-            var delta = -AppliedSpeedDifferential;
-            if (Math.Abs(delta) < 0.001f)
-                return;
-
-            foreach (var ammoId in AmmoDefinitionIds)
-            {
-                try
-                {
-                    var ammoDefinition = MyDefinitionManager.Static.GetAmmoDefinition(
-                        new MyDefinitionId(typeof(MyObjectBuilder_AmmoDefinition), ammoId));
-                    if (ammoDefinition != null)
-                        ammoDefinition.DesiredSpeed += delta;
-                }
-                catch
-                {
-                    // Ignore missing ammo definitions.
-                }
-            }
-
-            AppliedSpeedDifferential = 0f;
-        }
-        internal static void RefreshGroupsAfterConfigChanged()
-        {
-            var groups = new List<GroupComponent>(GroupDict.Values);
-            foreach (var group in groups)
-            {
-                if (group == null) continue;
-                group.OnConfigChanged();
-            }
         }
     }
 }
